[automa|fefactoring|models] - Refactored Automa models in to respective files & Also the AvatarViewImage (We can convert to using something more simple potentially!
