+++ Automa/AutomaApp.swift
@@ -5,8 +5,8 @@
 //  Created by Simon Ferns on 2024/04/23.
 //
 
-import SwiftUI
 import StripeCore
+import SwiftUI
 
 @main
 struct AutomaApp: App {
@@ -14,11 +14,12 @@ struct AutomaApp: App {
         WindowGroup {
             AppView()
                 .onOpenURL { incomingURL in
-                          let stripeHandled = StripeAPI.handleURLCallback(with: incomingURL)
-                          if (!stripeHandled) {
-                            // This was not a Stripe url – handle the URL normally as you would
-                          }
-                        }
+                    let stripeHandled = StripeAPI.handleURLCallback(with: incomingURL)
+                    if !stripeHandled {
+                        // This was not a Stripe url – handle the URL normally as you would
+                        print("Some other url \(incomingURL). Please handle this (This might require us to transform this into a 'Deeplink Provider'")
+                    }
+                }
         }
     }
 }
