+++ Automa/Components/GenericTitle.swift
@@ -7,6 +7,7 @@
 
 import Foundation
 import SwiftUI
+
 struct GenericTitle: View {
     let title: String
     let description: String
@@ -17,13 +18,12 @@ struct GenericTitle: View {
             .frame(maxWidth: .infinity, alignment: .leading)
             .padding(.leading, 26.0)
             .foregroundColor(Color("Primary"))
-        
-        Text(description)
-                        .font(.custom("Inter", size: 32))
-                        .fontWeight(.thin)
-                        .frame(maxWidth: .infinity, alignment: .leading)
-                        .padding(.leading, 25.0)
-                        .foregroundColor(Color("TextPrimary"))
 
+        Text(description)
+            .font(.custom("Inter", size: 32))
+            .fontWeight(.thin)
+            .frame(maxWidth: .infinity, alignment: .leading)
+            .padding(.leading, 25.0)
+            .foregroundColor(Color("TextPrimary"))
     }
 }
