+++ Automa/Models/Enums/TransferError.swift
@@ -1,0 +1,12 @@
+//
+//  TransferError.swift
+//  Automa
+//
+//  Created by Simon Ferns on 2024/05/13.
+//
+
+import Foundation
+
+enum TransferError: Error {
+    case importFailed
+}
