+++ Automa/ProfileView.swift
@@ -6,162 +6,160 @@
 //
 
 import Foundation
-import SwiftUI
-import Supabase
 import PhotosUI
+import Supabase
+import SwiftUI
 
 struct ProfileView: View {
-  @State var username = ""
-  @State var fullName = ""
-  @State var website = ""
-  @State var isLoading = false
- @State var imageSelection: PhotosPickerItem?
- @State var avatarImage: AvatarImage?
+    @State var username = ""
+    @State var fullName = ""
+    @State var website = ""
+    @State var isLoading = false
+    @State var imageSelection: PhotosPickerItem?
+    @State var avatarImage: AvatarImage?
 
-  var body: some View {
-    NavigationStack {
-        Form {
-            ZStack {
+    var body: some View {
+        NavigationStack {
+            Form {
+                ZStack {
                     if let avatarImage {
                         avatarImage.image.resizable().clipShape(/*@START_MENU_TOKEN@*/Circle()/*@END_MENU_TOKEN@*/).frame(height: 80)
                     } else {
                         Color.clear
                     }
-                
-                PhotosPicker(selection: $imageSelection, matching: .images) {
-                    Image(systemName: "pencil.circle.fill")
-                        .symbolRenderingMode(.multicolor)
-                        .font(.system(size: 30))
-                        .foregroundColor(.accentColor)
+
+                    PhotosPicker(selection: $imageSelection, matching: .images) {
+                        Image(systemName: "pencil.circle.fill")
+                            .symbolRenderingMode(.multicolor)
+                            .font(.system(size: 30))
+                            .foregroundColor(.accentColor)
+                    }
+                }
+
+                Section {
+                    TextField("Username", text: $username)
+                        .textContentType(.username)
+                        .textInputAutocapitalization(.never)
+                    TextField("Full name", text: $fullName)
+                        .textContentType(.name)
+                    TextField("Website", text: $website)
+                        .textContentType(.URL)
+                        .textInputAutocapitalization(.never)
+                }
+
+                Section {
+                    Button("Update profile") {
+                        updateProfileButtonTapped()
+                    }
+                    .bold()
+
+                    if isLoading {
+                        ProgressView()
+                    }
                 }
             }
-
-        Section {
-          TextField("Username", text: $username)
-            .textContentType(.username)
-            .textInputAutocapitalization(.never)
-          TextField("Full name", text: $fullName)
-            .textContentType(.name)
-          TextField("Website", text: $website)
-            .textContentType(.URL)
-            .textInputAutocapitalization(.never)
-        }
-
-        Section {
-          Button("Update profile") {
-            updateProfileButtonTapped()
-          }
-          .bold()
-
-          if isLoading {
-            ProgressView()
-          }
-        }
-            
-      }
-      .navigationTitle("Profile")
-      .toolbar(content: {
-        ToolbarItem {
-          Button("Sign out", role: .destructive) {
-            Task {
-              try? await supabase.auth.signOut()
+            .navigationTitle("Profile")
+            .toolbar(content: {
+                ToolbarItem {
+                    Button("Sign out", role: .destructive) {
+                        Task {
+                            try? await supabase.auth.signOut()
+                        }
+                    }
+                }
+            })
+            .onChange(of: imageSelection) { _, newValue in
+                guard let newValue else { return }
+                loadTransferable(from: newValue)
             }
-          }
         }
-      })
-      .onChange(of: imageSelection) { _, newValue in
-        guard let newValue else { return }
-        loadTransferable(from: newValue)
-      }
+        .task {
+            await getInitialProfile()
+        }
     }
-    .task {
-      await getInitialProfile()
+
+    func getInitialProfile() async {
+        do {
+            let currentUser = try await supabase.auth.session.user
+
+            let profile: Profile = try await supabase.database
+                .from("profiles")
+                .select()
+                .eq("id", value: currentUser.id)
+                .single()
+                .execute()
+                .value
+
+            username = profile.username ?? ""
+            fullName = profile.fullName ?? ""
+            website = profile.website ?? ""
+
+            if let avatarURL = profile.avatarURL, !avatarURL.isEmpty {
+                try await downloadImage(path: avatarURL)
+            }
+
+        } catch {
+            debugPrint(error)
+        }
     }
-      
-  }
 
-  func getInitialProfile() async {
-    do {
-      let currentUser = try await supabase.auth.session.user
+    func updateProfileButtonTapped() {
+        Task {
+            isLoading = true
+            defer { isLoading = false }
+            do {
+                let imageURL = try await uploadImage()
 
-      let profile: Profile = try await supabase.database
-        .from("profiles")
-        .select()
-        .eq("id", value: currentUser.id)
-        .single()
-        .execute()
-        .value
+                let currentUser = try await supabase.auth.session.user
 
-      username = profile.username ?? ""
-      fullName = profile.fullName ?? ""
-      website = profile.website ?? ""
+                let updatedProfile = Profile(
+                    username: username,
+                    fullName: fullName,
+                    website: website,
+                    avatarURL: imageURL,
+                    expiry_date: Date(),
+                    stripe_customer_id: ""
+                )
 
-      if let avatarURL = profile.avatarURL, !avatarURL.isEmpty {
-        try await downloadImage(path: avatarURL)
-      }
-
-    } catch {
-      debugPrint(error)
+                try await supabase.database
+                    .from("profiles")
+                    .update(updatedProfile)
+                    .eq("id", value: currentUser.id)
+                    .execute()
+            } catch {
+                debugPrint(error)
+            }
+        }
     }
-  }
 
-  func updateProfileButtonTapped() {
-    Task {
-      isLoading = true
-      defer { isLoading = false }
-      do {
-        let imageURL = try await uploadImage()
-
-        let currentUser = try await supabase.auth.session.user
-
-        let updatedProfile = Profile(
-          username: username,
-          fullName: fullName,
-          website: website,
-          avatarURL: imageURL,
-          expiry_date: Date(),
-          stripe_customer_id: ""
-        )
-
-        try await supabase.database
-          .from("profiles")
-          .update(updatedProfile)
-          .eq("id", value: currentUser.id)
-          .execute()
-      } catch {
-        debugPrint(error)
-      }
+    private func loadTransferable(from imageSelection: PhotosPickerItem) {
+        Task {
+            do {
+                avatarImage = try await imageSelection.loadTransferable(type: AvatarImage.self)
+            } catch {
+                debugPrint(error)
+            }
+        }
     }
-  }
 
-  private func loadTransferable(from imageSelection: PhotosPickerItem) {
-    Task {
-      do {
-        avatarImage = try await imageSelection.loadTransferable(type: AvatarImage.self)
-      } catch {
-        debugPrint(error)
-      }
+    private func downloadImage(path: String) async throws {
+        let data = try await supabase.storage.from("avatars").download(path: path)
+        avatarImage = AvatarImage(data: data)
     }
-  }
 
-  private func downloadImage(path: String) async throws {
-    let data = try await supabase.storage.from("avatars").download(path: path)
-    avatarImage = AvatarImage(data: data)
-  }
+    private func uploadImage() async throws -> String? {
+        guard let data = avatarImage?.data else { return nil }
 
-  private func uploadImage() async throws -> String? {
-    guard let data = avatarImage?.data else { return nil }
+        let filePath = "\(UUID().uuidString).jpeg"
 
-    let filePath = "\(UUID().uuidString).jpeg"
+        try await supabase.storage
+            .from("avatars")
+            .upload(
+                path: filePath,
+                file: data,
+                options: FileOptions(contentType: "image/jpeg")
+            )
 
-    try await supabase.storage
-      .from("avatars")
-      .upload(
-        path: filePath,
-        file: data,
-        options: FileOptions(contentType: "image/jpeg")
-      )
-
-    return filePath
-  }
+        return filePath
+    }
 }
