+++ Automa/Screens/AuthScreen/Alerts/EmailResentAlert.swift
@@ -5,8 +5,8 @@
 //  Created by Simon Ferns on 2024/05/06.
 //
 
-import Foundation
 import AlertToast
+import Foundation
 
 func EmailResentAlert() -> AlertToast {
     return AlertToast(type: .regular, title: "Email Sent Again!")
