+++ Automa/Screens/AuthScreen/AuthScreen.swift
@@ -1,5 +1,5 @@
 //
-//  AuthView.swift
+//  AuthScreen.swift
 //  Automa
 //
 //  Created by Simon Ferns on 2024/05/03.
@@ -22,7 +22,7 @@ struct AuthScreen: View {
             VStack(spacing: 18) {
                 VStack(spacing: 10) {
                     GenericTitle(title: "Lets sign you in.", description: "Welcome back. \nLets get automating!")
-               }
+                }
 
                 Spacer()
