+++ Automa/Screens/AuthScreen/AuthScreenModel.swift
@@ -1,5 +1,5 @@
 //
-//  AuthViewModel.swift
+//  AuthScreenModel.swift
 //  Automa
 //
 //  Created by Simon Ferns on 2024/05/06.
