+++ Automa/Screens/PlanPaymentScreen/PlanPaymentScreen.swift
@@ -1,5 +1,5 @@
 //
-//  CheckoutView.swift
+//  PlanPaymentScreen.swift
 //  Automa
 //
 //  Created by Simon Ferns on 2024/05/03.
@@ -8,7 +8,6 @@
 import Foundation
 import StripePaymentSheet
 import SwiftUI
-
 
 struct PlanPaymentScreen: View {
     @ObservedObject var model = PaymentSheetModel()
@@ -51,7 +50,7 @@ struct PlanPaymentScreen: View {
 
                                 ForEach(plans[1...]) { plan in
                                     PaymentButton(model: model, setPaymentBool: setPaymentBool, showErrorDialog: showErrorDialog, paymentType: plan.id, title: plan.name, price: plan.price_rep, backgroundColor: UInt(plan.hex_color_int))
-                                    }
+                                }
                             }
 
                         }.tint(.white)
