+++ AutomaTests/AutomaTests.swift
@@ -5,11 +5,10 @@
 //  Created by Simon Ferns on 2024/04/23.
 //
 
-import XCTest
 @testable import Automa
+import XCTest
 
 final class AutomaTests: XCTestCase {
-
     override func setUpWithError() throws {
         // Put setup code here. This method is called before the invocation of each test method in the class.
     }
@@ -28,9 +27,8 @@ final class AutomaTests: XCTestCase {
 
     func testPerformanceExample() throws {
         // This is an example of a performance test case.
-        self.measure {
+        measure {
             // Put the code you want to measure the time of here.
         }
     }
-
 }
