+++ AutomaUITests/AutomaUITests.swift
@@ -8,7 +8,6 @@
 import XCTest
 
 final class AutomaUITests: XCTestCase {
-
     override func setUpWithError() throws {
         // Put setup code here. This method is called before the invocation of each test method in the class.
