+++ AutomaUITests/AutomaUITestsLaunchTests.swift
@@ -8,7 +8,6 @@
 import XCTest
 
 final class AutomaUITestsLaunchTests: XCTestCase {
-
     override class var runsForEachTargetApplicationUIConfiguration: Bool {
         true
     }
